+++ ApexEngineV2/rendering/mesh.cpp
@@ -48,7 +48,7 @@ void Mesh::SetVertices(const std::vector<Vertex> &verts)
     is_uploaded = false;
 }
 
-void Mesh::SetVertices(const std::vector<Vertex> &verts, const std::vector<uint32_t> &ind)
+void Mesh::SetVertices(const std::vector<Vertex> &verts, const std::vector<size_t> &ind)
 {
     vertices = verts;
     indices = ind;
